EBS source working for single EBS volume case
Because of differences in how device mapping happens between when
brenda was built and today, we can only theoretically support 11 EBS
volumes.  Additionally, we have to hack in a linking of devices on the
workers as part of the startup script to allow Brenda to be able to
find the volume.  For now we’ll limit to only allowing a single EBS
volume to be used as a source and leave the 11 EBS volume case for
later.